backend/requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==26.1.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
blis==1.3.3
build==1.4.0
catalogue==2.0.10
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
chromadb==1.5.5
click==8.3.1
cloudpathlib==0.23.0
colorama==0.4.6
confection==0.1.5
cryptography==46.0.5
cymem==2.0.13
dataclasses-json==0.6.7
docx2pdf==0.1.8
durationpy==0.10
emoji==2.15.0
fastapi==0.135.1
filelock==3.25.2
filetype==1.2.0
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.2.0
googleapis-common-protos==1.73.0
greenlet==3.3.2
grpcio==1.78.0
h11==0.16.0
hf-xet==1.4.2
html5lib==1.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.7.2
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
installer==0.7.0
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kubernetes==35.0.0
langchain==1.2.13
langchain-chroma==1.1.0
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.20
langchain-ollama==1.0.1
langchain-text-splitters==1.1.1
langdetect==1.0.9
langgraph==1.1.3
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.7.22
llvmlite==0.46.0
lxml==6.0.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
mmh3==5.2.1
mpmath==1.3.0
multidict==6.7.1
murmurhash==1.0.15
mypy_extensions==1.1.0
numba==0.64.0
numpy==2.4.3
oauthlib==3.3.1
olefile==0.47
ollama==0.6.1
onnxruntime==1.24.4
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
orjson==3.11.7
ormsgpack==1.12.2
overrides==7.7.0
packaging==26.0
pandas==3.0.1
pillow==12.1.1
preshed==3.0.12
propcache==0.4.1
protobuf==6.33.6
psutil==7.2.2
psycopg2-binary==2.9.11
pybase64==1.4.3
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
PyMuPDF==1.27.2.2
pypdf==6.9.1
pypdfium2==5.6.0
PyPika==0.51.1
pyproject_hooks==1.2.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.2
python-iso639==2026.1.31
python-magic==0.4.27
python-multipart==0.0.22
python-oxmsg==0.0.2
python-pptx==1.0.2
pywin32==311
PyYAML==6.0.3
RapidFuzz==3.14.3
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.3.3
rpds-py==0.30.0
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
smart_open==7.5.1
soupsieve==2.8.3
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.48
srsly==2.5.2
starlette==0.52.1
sympy==1.14.0
tenacity==9.1.4
thinc==8.3.11
tokenizers==0.22.2
tqdm==4.67.3
typer==0.24.1
typer-slim==0.24.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
unstructured==0.21.5
unstructured-client==0.42.10
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.42.0
wasabi==1.1.3
watchfiles==1.1.1
weasel==0.4.3
webencodings==0.5.1
websocket-client==1.9.0
websockets==16.0
wrapt==1.17.3
xlsxwriter==3.2.9
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.0
zstandard==0.25.0